WIP Volumes

+++ bismuth.core.tests/src/containers/darray_tests.c
@@ -979,7 +979,7 @@ static u8 darray_all_iterator_tests(void)
         // Try reverse/backward iteration on an empty array
         it = arr.rbegin(&arr.base);
         expect_should_be(&arr.base, it.arr);
-        expect_should_be(arr.base.length - 1, it.pos);
+        expect_should_be(arr.base.length - 1, (u32)it.pos);
         expect_should_be(-1, it.dir);
         loop_count = 0;
         for (; !it.end(&it); it.next(&it))
@@ -1016,7 +1016,7 @@ static u8 darray_all_iterator_tests(void)
         // Try reverse/backward iteration
         it = arr.rbegin(&arr.base);
         expect_should_be(&arr.base, it.arr);
-        expect_should_be(arr.base.length - 1, it.pos);
+        expect_should_be(arr.base.length - 1, (u32)it.pos);
         expect_should_be(-1, it.dir);
         loop_count = 0;
         for (; !it.end(&it); it.next(&it))
@@ -1064,7 +1064,7 @@ static u8 darray_all_iterator_tests(void)
         // Try reverse/backward iteration
         it = arr.rbegin(&arr.base);
         expect_should_be(&arr.base, it.arr);
-        expect_should_be(arr.base.length - 1, it.pos);
+        expect_should_be(arr.base.length - 1, (u32)it.pos);
         expect_should_be(-1, it.dir);
         loop_count = 0;
         for (; !it.end(&it); it.next(&it))
@@ -1124,7 +1124,7 @@ static u8 darray_all_iterator_tests(void)
         // Try reverse/backward iteration
         it = arr.rbegin(&arr.base);
         expect_should_be(&arr.base, it.arr);
-        expect_should_be(arr.base.length - 1, it.pos);
+        expect_should_be(arr.base.length - 1, (u32)it.pos);
         expect_should_be(-1, it.dir);
         loop_count = 0;
         for (; !it.end(&it); it.next(&it))
